--- src/pre_tables.c
@@ -1,20 +1,0 @@
-//
-// Created by erena on 18.07.2024.
-//
-
-#include "pre_tables.h"
-
-void initAll() {
-    initLeaperAttacks();
-    initMagicNumbers();
-    initSlidersAttacks(bishop);
-    initSlidersAttacks(rook);
-    // init random keys for tranposition table
-    //initRandomKeys();
-    // clear hash table
-    clearHashTable();
-    // init mask
-    initEvaluationMasks();
-    // init Late Move Reduction Table
-    initializeLMRTable();
-}
