--- src/pre_tables.h
@@ -1,11 +1,0 @@
-//
-// Created by erena on 18.07.2024.
-//
-#include "move.h"
-#include "search.h"
-#include "table.h"
-#include "magic.h"
-#include "magic.c"
-#include "zobrist.h"
-
-extern void initAll();
